@@ -1,53 +1,72 @@
 {
-  "name": "sb-admin-angular4-bootstrap4",
-  "version": "1.0.0",
+  "name": "SB-Admin-BS4-Angular-5",
+  "version": "4.0.0",
   "license": "MIT",
   "scripts": {
     "ng": "ng",
-    "start": "ng serve --ec=true",
+    "postinstall": "ng build --aot -prod",
+    "start": "node server.js",
     "build": "ng build --prod",
+    "gitbuild": "ng build --prod --base /start-angular/SB-Admin-BS4-Angular-5/master/dist/",
     "test": "ng test",
     "lint": "ng lint",
     "e2e": "ng e2e"
   },
   "private": true,
   "dependencies": {
-    "@angular/common": "^4.0.0",
-    "@angular/compiler": "^4.0.0",
-    "@angular/core": "^4.0.0",
-    "@angular/forms": "^4.0.0",
-    "@angular/http": "^4.0.0",
-    "@angular/platform-browser": "^4.0.0",
-    "@angular/platform-browser-dynamic": "^4.0.0",
-    "@angular/router": "^4.0.0",
-    "@ng-bootstrap/ng-bootstrap": "^1.0.0-alpha.25",
-    "@ngx-translate/core": "^6.0.1",
-    "@ngx-translate/http-loader": "0.0.3",
-    "core-js": "^2.4.1",
+    "@agm/core": "^1.0.0-beta.2",
+    "@angular/animations": "^5.0.1",
+    "@angular/cli": "^1.6.7",
+    "@angular/common": "^5.0.1",
+    "@angular/compiler": "^5.0.1",
+    "@angular/compiler-cli": "^5.2.3",
+    "@angular/core": "^5.0.1",
+    "@angular/forms": "^5.0.1",
+    "@angular/http": "^5.0.1",
+    "@angular/platform-browser": "^5.0.1",
+    "@angular/platform-browser-dynamic": "^5.0.1",
+    "@angular/router": "^5.0.1",
+    "@ng-bootstrap/ng-bootstrap": "^1.0.0-beta.5",
+    "@ngx-translate/core": "^8.0.0",
+    "@ngx-translate/http-loader": "^2.0.0",
+    "@types/googlemaps": "^3.30.2",
+    "chart.js": "^2.7.1",
+    "core-js": "^2.5.1",
+    "express": "^4.16.2",
     "font-awesome": "^4.7.0",
-    "ionicons": "^3.0.0",
-    "ng2-charts": "^1.5.0",
-    "rxjs": "^5.1.0",
-    "zone.js": "^0.8.4"
+    "ng2-charts": "^1.6.0",
+    "ngx-bootstrap": "^2.0.2",
+    "node-sass": "^4.7.2",
+    "primeng": "^5.0.2",
+    "rxjs": "^5.5.2",
+    "yarn": "^1.3.2",
+    "zone.js": "^0.8.18"
   },
   "devDependencies": {
-    "@angular/cli": "1.0.3",
-    "@angular/compiler-cli": "^4.0.0",
-    "@types/jasmine": "2.5.38",
-    "@types/node": "~6.0.60",
-    "codelyzer": "~2.0.0",
-    "jasmine-core": "~2.5.2",
-    "jasmine-spec-reporter": "~3.2.0",
-    "karma": "~1.4.1",
-    "karma-chrome-launcher": "~2.0.0",
+    "@angular/cli": "^1.6.7",
+    "@angular/compiler-cli": "^5.2.3",
+    "@angular/language-service": "^5.0.1",
+    "@types/jasmine": "~2.6.3",
+    "@types/jasminewd2": "~2.0.3",
+    "@types/node": "~8.0.51",
+    "codelyzer": "~4.0.1",
+    "enhanced-resolve": "^3.3.0",
+    "jasmine-core": "~2.8.0",
+    "jasmine-spec-reporter": "~4.2.1",
+    "karma": "~1.7.0",
+    "karma-chrome-launcher": "~2.2.0",
     "karma-cli": "~1.0.1",
-    "karma-coverage-istanbul-reporter": "^0.2.0",
+    "karma-coverage-istanbul-reporter": "^1.2.1",
     "karma-jasmine": "~1.1.0",
     "karma-jasmine-html-reporter": "^0.2.2",
-    "karma-phantomjs-launcher": "^1.0.4",
-    "protractor": "~5.1.0",
-    "ts-node": "~2.0.0",
-    "tslint": "~4.5.0",
-    "typescript": "~2.2.0"
+    "protractor": "~5.1.2",
+    "ts-node": "~3.3.0",
+    "tslint": "~5.7.0",
+    "typescript": "^2.6.2",
+    "webpack": "^3.8.1"
+  },
+  "engines": {
+    "node": "6.11.4",
+    "npm": "5.5.1"
   }
 }
