@@ -1,14 +1,14 @@
-import { CliStablePage } from './app.po';
+import { AppPage } from './app.po';
 
-describe('cli-stable App', () => {
-  let page: CliStablePage;
+describe('test-ng4 App', () => {
+  let page: AppPage;
 
   beforeEach(() => {
-    page = new CliStablePage();
+    page = new AppPage();
   });
 
-  it('should display message saying app works', () => {
+  it('should display welcome message', () => {
     page.navigateTo();
-    expect(page.getParagraphText()).toEqual('app works!');
+    expect(page.getParagraphText()).toEqual('Welcome to app!');
   });
 });
