@@ -1,15 +1,13 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
-import { AuthGuard } from './shared';
+import { AppComponent } from './app.component';
 
 const routes: Routes = [
-    {
-        path: '',
-        loadChildren: './layout/layout.module#LayoutModule',
-        canActivate: [AuthGuard]
-    },
+    { path: '', loadChildren: './layout/layout.module#LayoutModule'},
     { path: 'login', loadChildren: './login/login.module#LoginModule' },
     { path: 'signup', loadChildren: './signup/signup.module#SignupModule' },
+    { path: 'error', loadChildren: './server-error/server-error.module#ServerErrorModule' },
+    { path: 'access-denied', loadChildren: './access-denied/access-denied.module#AccessDeniedModule' },
     { path: 'not-found', loadChildren: './not-found/not-found.module#NotFoundModule' },
     { path: '**', redirectTo: 'not-found' }
 ];
@@ -18,4 +16,4 @@ const routes: Routes = [
     imports: [RouterModule.forRoot(routes)],
     exports: [RouterModule]
 })
-export class AppRoutingModule { }
+export class AppRoutingModule {}
