@@ -1,21 +1,14 @@
-import { Component } from '@angular/core';
-import { Router } from '@angular/router';
-import { TranslateService } from '@ngx-translate/core';
+import { Component, OnInit } from '@angular/core';
 
 @Component({
     selector: 'app-root',
     templateUrl: './app.component.html',
     styleUrls: ['./app.component.scss']
 })
-export class AppComponent {
-
-    constructor(private translate: TranslateService) {
-        translate.addLangs(['en', 'fr', 'ur']);
-        translate.setDefaultLang('en');
-
-        const browserLang = translate.getBrowserLang();
-        translate.use(browserLang.match(/en|fr|ur/) ? browserLang : 'en');
+export class AppComponent implements OnInit {
+    constructor() {
     }
 
+    ngOnInit() {
+    }
 }
-
