@@ -1,11 +1,14 @@
 import { Component, OnInit } from '@angular/core';
+import { routerTransition } from '../../router.animations';
 
 @Component({
     selector: 'app-blank-page',
     templateUrl: './blank-page.component.html',
-    styleUrls: ['./blank-page.component.scss']
+    styleUrls: ['./blank-page.component.scss'],
+    animations: [routerTransition()]
 })
 export class BlankPageComponent implements OnInit {
-    constructor() { }
-    ngOnInit() { }
+    constructor() {}
+
+    ngOnInit() {}
 }
