@@ -1,9 +1,12 @@
 import { Component, OnInit } from '@angular/core';
-import {NgbModal, ModalDismissReasons} from '@ng-bootstrap/ng-bootstrap';
 
 @Component({
     selector: 'app-bs-component',
     templateUrl: './bs-component.component.html',
     styleUrls: ['./bs-component.component.scss']
 })
-export class BsComponentComponent { }
+export class BsComponentComponent implements OnInit {
+    constructor() {}
+
+    ngOnInit() {}
+}
