@@ -3,7 +3,6 @@ import { CommonModule } from '@angular/common';
 import { NgbModule } from '@ng-bootstrap/ng-bootstrap';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 
-// relative import
 import { BsComponentRoutingModule } from './bs-component-routing.module';
 import { BsComponentComponent } from './bs-component.component';
 import {
@@ -18,20 +17,21 @@ import {
     ProgressbarComponent,
     TabsComponent,
     TooltipComponent,
-    TimepickerComponent,
+    TimepickerComponent    
 } from './components';
 import { PageHeaderModule } from '../../shared';
 
 @NgModule({
     imports: [
         CommonModule,
+        BsComponentRoutingModule,
         FormsModule,
         ReactiveFormsModule,
         NgbModule.forRoot(),
-        PageHeaderModule,
-        BsComponentRoutingModule,
+        PageHeaderModule
     ],
     declarations: [
+        BsComponentComponent,
         BsComponentComponent,
         ButtonsComponent,
         AlertComponent,
@@ -47,4 +47,4 @@ import { PageHeaderModule } from '../../shared';
         TimepickerComponent
     ]
 })
-export class BsComponentModule { }
+export class BsComponentModule {}
