@@ -1,11 +1,14 @@
 import { Component, OnInit } from '@angular/core';
+import { routerTransition } from '../../router.animations';
 
 @Component({
     selector: 'app-bs-element',
     templateUrl: './bs-element.component.html',
-    styleUrls: ['./bs-element.component.scss']
+    styleUrls: ['./bs-element.component.scss'],
+    animations: [routerTransition()]
 })
 export class BsElementComponent implements OnInit {
-    constructor() { }
-    ngOnInit() { }
+    constructor() {}
+
+    ngOnInit() {}
 }
