@@ -1,14 +1,16 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
-
 import { ChartsComponent } from './charts.component';
 
 const routes: Routes = [
-    { path: '', component: ChartsComponent }
+    {
+        path: '',
+        component: ChartsComponent
+    }
 ];
 
 @NgModule({
     imports: [RouterModule.forChild(routes)],
     exports: [RouterModule]
 })
-export class ChartsRoutingModule { }
+export class ChartsRoutingModule {}
