@@ -1,34 +1,26 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-import { RouterTestingModule } from '@angular/router/testing';
-import { ChartsModule as Ng2Charts } from 'ng2-charts';
-
-import { PageHeaderModule } from './../../shared';
 
 import { ChartsComponent } from './charts.component';
 
 describe('ChartsComponent', () => {
-  let component: ChartsComponent;
-  let fixture: ComponentFixture<ChartsComponent>;
+    let component: ChartsComponent;
+    let fixture: ComponentFixture<ChartsComponent>;
 
-  beforeEach(async(() => {
-    TestBed.configureTestingModule({
-    imports: [
-      RouterTestingModule,
-      Ng2Charts,
-      PageHeaderModule
-    ],
-      declarations: [ ChartsComponent ]
-    })
-    .compileComponents();
-  }));
+    beforeEach(
+        async(() => {
+            TestBed.configureTestingModule({
+                declarations: [ChartsComponent]
+            }).compileComponents();
+        })
+    );
 
-  beforeEach(() => {
-    fixture = TestBed.createComponent(ChartsComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
+    beforeEach(() => {
+        fixture = TestBed.createComponent(ChartsComponent);
+        component = fixture.componentInstance;
+        fixture.detectChanges();
+    });
 
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
+    it('should create', () => {
+        expect(component).toBeTruthy();
+    });
 });
