@@ -1,9 +1,11 @@
 import { Component, OnInit } from '@angular/core';
+import { routerTransition } from '../../router.animations';
 
 @Component({
     selector: 'app-charts',
     templateUrl: './charts.component.html',
-    styleUrls: ['./charts.component.scss']
+    styleUrls: ['./charts.component.scss'],
+    animations: [routerTransition()]
 })
 export class ChartsComponent implements OnInit {
     // bar chart
@@ -11,70 +13,116 @@ export class ChartsComponent implements OnInit {
         scaleShowVerticalLines: false,
         responsive: true
     };
-    public barChartLabels: string[] = ['2006', '2007', '2008', '2009', '2010', '2011', '2012'];
+    public barChartLabels: string[] = [
+        '2006',
+        '2007',
+        '2008',
+        '2009',
+        '2010',
+        '2011',
+        '2012'
+    ];
     public barChartType: string = 'bar';
     public barChartLegend: boolean = true;
 
     public barChartData: any[] = [
-    {data: [65, 59, 80, 81, 56, 55, 40], label: 'Series A'},
-    {data: [28, 48, 40, 19, 86, 27, 90], label: 'Series B'}
+        { data: [65, 59, 80, 81, 56, 55, 40], label: 'Series A' },
+        { data: [28, 48, 40, 19, 86, 27, 90], label: 'Series B' }
     ];
+
     // Doughnut
-    public doughnutChartLabels: string[] = ['Download Sales', 'In-Store Sales', 'Mail-Order Sales'];
+    public doughnutChartLabels: string[] = [
+        'Download Sales',
+        'In-Store Sales',
+        'Mail-Order Sales'
+    ];
     public doughnutChartData: number[] = [350, 450, 100];
     public doughnutChartType: string = 'doughnut';
+
     // Radar
-    public radarChartLabels: string[] = ['Eating', 'Drinking', 'Sleeping', 'Designing', 'Coding', 'Cycling', 'Running'];
+    public radarChartLabels: string[] = [
+        'Eating',
+        'Drinking',
+        'Sleeping',
+        'Designing',
+        'Coding',
+        'Cycling',
+        'Running'
+    ];
     public radarChartData: any = [
-        {data: [65, 59, 90, 81, 56, 55, 40], label: 'Series A'},
-        {data: [28, 48, 40, 19, 96, 27, 100], label: 'Series B'}
+        { data: [65, 59, 90, 81, 56, 55, 40], label: 'Series A' },
+        { data: [28, 48, 40, 19, 96, 27, 100], label: 'Series B' }
     ];
     public radarChartType: string = 'radar';
-    // Pie
-    public pieChartLabels:string[] = ['Download Sales', 'In-Store Sales', 'Mail Sales'];
-    public pieChartData:number[] = [300, 500, 100];
-    public pieChartType:string = 'pie';
-    // PolarArea
-    public polarAreaChartLabels:string[] = ['Download Sales', 'In-Store Sales', 'Mail Sales', 'Telesales', 'Corporate Sales'];
-    public polarAreaChartData:number[] = [300, 500, 100, 40, 120];
-    public polarAreaLegend:boolean = true;
 
-    public polarAreaChartType:string = 'polarArea';
+    // Pie
+    public pieChartLabels: string[] = [
+        'Download Sales',
+        'In-Store Sales',
+        'Mail Sales'
+    ];
+    public pieChartData: number[] = [300, 500, 100];
+    public pieChartType: string = 'pie';
+
+    // PolarArea
+    public polarAreaChartLabels: string[] = [
+        'Download Sales',
+        'In-Store Sales',
+        'Mail Sales',
+        'Telesales',
+        'Corporate Sales'
+    ];
+    public polarAreaChartData: number[] = [300, 500, 100, 40, 120];
+    public polarAreaLegend: boolean = true;
+
+    public polarAreaChartType: string = 'polarArea';
+
     // lineChart
     public lineChartData: Array<any> = [
-        {data: [65, 59, 80, 81, 56, 55, 40], label: 'Series A'},
-        {data: [28, 48, 40, 19, 86, 27, 90], label: 'Series B'},
-        {data: [18, 48, 77, 9, 100, 27, 40], label: 'Series C'}
+        { data: [65, 59, 80, 81, 56, 55, 40], label: 'Series A' },
+        { data: [28, 48, 40, 19, 86, 27, 90], label: 'Series B' },
+        { data: [18, 48, 77, 9, 100, 27, 40], label: 'Series C' }
     ];
-    public lineChartLabels: Array<any> = ['January', 'February', 'March', 'April', 'May', 'June', 'July'];
+    public lineChartLabels: Array<any> = [
+        'January',
+        'February',
+        'March',
+        'April',
+        'May',
+        'June',
+        'July'
+    ];
     public lineChartOptions: any = {
         responsive: true
     };
     public lineChartColors: Array<any> = [
-    { // grey
-        backgroundColor: 'rgba(148,159,177,0.2)',
-        borderColor: 'rgba(148,159,177,1)',
-        pointBackgroundColor: 'rgba(148,159,177,1)',
-        pointBorderColor: '#fff',
-        pointHoverBackgroundColor: '#fff',
-        pointHoverBorderColor: 'rgba(148,159,177,0.8)'
-    },
-    { // dark grey
-        backgroundColor: 'rgba(77,83,96,0.2)',
-        borderColor: 'rgba(77,83,96,1)',
-        pointBackgroundColor: 'rgba(77,83,96,1)',
-        pointBorderColor: '#fff',
-        pointHoverBackgroundColor: '#fff',
-        pointHoverBorderColor: 'rgba(77,83,96,1)'
-    },
-    { // grey
-        backgroundColor: 'rgba(148,159,177,0.2)',
-        borderColor: 'rgba(148,159,177,1)',
-        pointBackgroundColor: 'rgba(148,159,177,1)',
-        pointBorderColor: '#fff',
-        pointHoverBackgroundColor: '#fff',
-        pointHoverBorderColor: 'rgba(148,159,177,0.8)'
-    }
+        {
+            // grey
+            backgroundColor: 'rgba(148,159,177,0.2)',
+            borderColor: 'rgba(148,159,177,1)',
+            pointBackgroundColor: 'rgba(148,159,177,1)',
+            pointBorderColor: '#fff',
+            pointHoverBackgroundColor: '#fff',
+            pointHoverBorderColor: 'rgba(148,159,177,0.8)'
+        },
+        {
+            // dark grey
+            backgroundColor: 'rgba(77,83,96,0.2)',
+            borderColor: 'rgba(77,83,96,1)',
+            pointBackgroundColor: 'rgba(77,83,96,1)',
+            pointBorderColor: '#fff',
+            pointHoverBackgroundColor: '#fff',
+            pointHoverBorderColor: 'rgba(77,83,96,1)'
+        },
+        {
+            // grey
+            backgroundColor: 'rgba(148,159,177,0.2)',
+            borderColor: 'rgba(148,159,177,1)',
+            pointBackgroundColor: 'rgba(148,159,177,1)',
+            pointBorderColor: '#fff',
+            pointHoverBackgroundColor: '#fff',
+            pointHoverBorderColor: 'rgba(148,159,177,0.8)'
+        }
     ];
     public lineChartLegend: boolean = true;
     public lineChartType: string = 'line';
@@ -94,21 +142,23 @@ export class ChartsComponent implements OnInit {
             Math.round(Math.random() * 100),
             59,
             80,
-            (Math.random() * 100),
+            Math.random() * 100,
             56,
-            (Math.random() * 100),
+            Math.random() * 100,
             40
         ];
         const clone = JSON.parse(JSON.stringify(this.barChartData));
         clone[0].data = data;
         this.barChartData = clone;
         /**
-        * (My guess), for Angular to recognize the change in the dataset
-        * it has to change the dataset variable directly,
-        * so one way around it, is to clone the data, change it and then
-        * assign it;
-        */
+         * (My guess), for Angular to recognize the change in the dataset
+         * it has to change the dataset variable directly,
+         * so one way around it, is to clone the data, change it and then
+         * assign it;
+         */
     }
-    constructor() { }
-    ngOnInit() { }
+
+    constructor() {}
+
+    ngOnInit() {}
 }
