@@ -7,12 +7,7 @@ import { ChartsComponent } from './charts.component';
 import { PageHeaderModule } from '../../shared';
 
 @NgModule({
-    imports: [
-        CommonModule,
-        Ng2Charts,
-        ChartsRoutingModule,
-        PageHeaderModule
-    ],
+    imports: [CommonModule, Ng2Charts, ChartsRoutingModule, PageHeaderModule],
     declarations: [ChartsComponent]
 })
-export class ChartsModule { }
+export class ChartsModule {}
