@@ -1,12 +1,5 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-import { NgbCarouselModule, NgbAlertModule } from '@ng-bootstrap/ng-bootstrap';
 
-import {
-  TimelineComponent,
-  NotificationComponent,
-  ChatComponent
-} from './components';
-import { StatModule } from '../../shared';
 import { DashboardComponent } from './dashboard.component';
 
 describe('DashboardComponent', () => {
@@ -15,17 +8,7 @@ describe('DashboardComponent', () => {
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
-    imports: [
-        NgbCarouselModule.forRoot(),
-        NgbAlertModule.forRoot(),
-        StatModule,
-    ],
-      declarations: [
-        DashboardComponent,
-        TimelineComponent,
-        NotificationComponent,
-        ChatComponent
-      ]
+      declarations: [ DashboardComponent ]
     })
     .compileComponents();
   }));
