@@ -1,14 +1,16 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
-
 import { FormComponent } from './form.component';
 
 const routes: Routes = [
-    { path: '', component: FormComponent }
+    {
+        path: '', component: FormComponent
+    }
 ];
 
 @NgModule({
-  imports: [RouterModule.forChild(routes)],
-  exports: [RouterModule]
+    imports: [RouterModule.forChild(routes)],
+    exports: [RouterModule]
 })
-export class FormRoutingModule { }
+export class FormRoutingModule {
+}
