@@ -1,11 +1,14 @@
 import { Component, OnInit } from '@angular/core';
+import { routerTransition } from '../../router.animations';
 
 @Component({
     selector: 'app-form',
     templateUrl: './form.component.html',
-    styleUrls: ['./form.component.scss']
+    styleUrls: ['./form.component.scss'],
+    animations: [routerTransition()]
 })
 export class FormComponent implements OnInit {
-    constructor() { }
+    constructor() {}
+
     ngOnInit() {}
 }
