@@ -1,5 +1,4 @@
 import { Component, OnInit } from '@angular/core';
-import { Router } from '@angular/router';
 
 @Component({
     selector: 'app-layout',
@@ -7,13 +6,7 @@ import { Router } from '@angular/router';
     styleUrls: ['./layout.component.scss']
 })
 export class LayoutComponent implements OnInit {
+    constructor() {}
 
-    constructor(public router: Router) { }
-
-    ngOnInit() {
-        if (this.router.url === '/') {
-            this.router.navigate(['/dashboard']);
-        }
-    }
-
+    ngOnInit() {}
 }
