@@ -1,23 +1,20 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { NgbDropdownModule } from '@ng-bootstrap/ng-bootstrap';
 import { TranslateModule } from '@ngx-translate/core';
+import { NgbDropdownModule } from '@ng-bootstrap/ng-bootstrap';
 
 import { LayoutRoutingModule } from './layout-routing.module';
 import { LayoutComponent } from './layout.component';
-import { HeaderComponent, SidebarComponent } from '../shared';
+import { SidebarComponent } from './components/sidebar/sidebar.component';
+import { HeaderComponent } from './components/header/header.component';
 
 @NgModule({
     imports: [
         CommonModule,
-        NgbDropdownModule.forRoot(),
         LayoutRoutingModule,
-        TranslateModule
+        TranslateModule,
+        NgbDropdownModule.forRoot()
     ],
-    declarations: [
-        LayoutComponent,
-        HeaderComponent,
-        SidebarComponent,
-    ]
+    declarations: [LayoutComponent, SidebarComponent, HeaderComponent]
 })
-export class LayoutModule { }
+export class LayoutModule {}
