@@ -3,11 +3,14 @@ import { Routes, RouterModule } from '@angular/router';
 import { TablesComponent } from './tables.component';
 
 const routes: Routes = [
-    { path: '', component: TablesComponent }
+    {
+        path: '', component: TablesComponent
+    }
 ];
 
 @NgModule({
     imports: [RouterModule.forChild(routes)],
     exports: [RouterModule]
 })
-export class TablesRoutingModule { }
+export class TablesRoutingModule {
+}
