@@ -1,11 +1,14 @@
 import { Component, OnInit } from '@angular/core';
+import { routerTransition } from '../../router.animations';
 
 @Component({
     selector: 'app-tables',
     templateUrl: './tables.component.html',
-    styleUrls: ['./tables.component.scss']
+    styleUrls: ['./tables.component.scss'],
+    animations: [routerTransition()]
 })
 export class TablesComponent implements OnInit {
-    constructor() { }
-    ngOnInit() { }
+    constructor() {}
+
+    ngOnInit() {}
 }
