@@ -1,16 +1,12 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 
-import { TablesComponent } from './tables.component';
 import { TablesRoutingModule } from './tables-routing.module';
+import { TablesComponent } from './tables.component';
 import { PageHeaderModule } from './../../shared';
 
 @NgModule({
-    imports: [
-        CommonModule,
-        TablesRoutingModule,
-        PageHeaderModule
-    ],
+    imports: [CommonModule, TablesRoutingModule, PageHeaderModule],
     declarations: [TablesComponent]
 })
-export class TablesModule { }
+export class TablesModule {}
