@@ -3,11 +3,14 @@ import { Routes, RouterModule } from '@angular/router';
 import { LoginComponent } from './login.component';
 
 const routes: Routes = [
-    { path: '', component: LoginComponent }
+    {
+        path: '',
+        component: LoginComponent
+    }
 ];
 
 @NgModule({
-  imports: [RouterModule.forChild(routes)],
-  exports: [RouterModule]
+    imports: [RouterModule.forChild(routes)],
+    exports: [RouterModule]
 })
-export class LoginRoutingModule { }
+export class LoginRoutingModule {}
