@@ -1,19 +1,69 @@
 import { Component, OnInit } from '@angular/core';
-import { Router } from '@angular/router';
+import { Router, ActivatedRoute } from '@angular/router';
+import { routerTransition } from '../router.animations';
+
+import { AlertService, AuthenticationService, UserService } from '../login/_services/index';
 
 @Component({
     selector: 'app-login',
     templateUrl: './login.component.html',
-    styleUrls: ['./login.component.scss']
+    styleUrls: ['./login.component.scss'],
+    animations: [routerTransition()]
 })
 export class LoginComponent implements OnInit {
+    model: any = {};
+    modelr: any =  {};
+    loading = false;
+    returnUrl: string;
 
-    constructor(public router: Router) { }
+    constructor(
+        private route: ActivatedRoute,
+        private router: Router,
+        private userService: UserService,
+        private authenticationService: AuthenticationService,
+        private alertService: AlertService) {}
 
-    ngOnInit() { }
-
-    onLoggedin() {
-        localStorage.setItem('isLoggedin', 'true');
+    ngOnInit() {
+        // reset login status
+        this.authenticationService.logout();
+        
+                // get return url from route parameters or default to '/'
+                this.returnUrl = this.route.snapshot.queryParams['returnUrl'] || '/';
+                console.log("qui...");
+                console.log("loading = "+this.loading);
     }
 
+    onLoggedin() {
+        console.log("pure qui..."+this.model.username+this.model.password);
+        this.loading = true;
+        console.log("loading = "+this.loading);
+        //creazione utente fittizio.
+        localStorage.removeItem('users');
+        this.modelr.username = "vvf";
+        this.modelr.password = "vvf";
+        this.userService.create(this.modelr)
+        .subscribe(
+            data => {
+             //   this.alertService.success('Registration successful', true);
+             //   this.router.navigate(['/login']);
+            },
+            error => {
+             //   this.alertService.error(error);
+                this.loading = false;
+            });
+
+        this.authenticationService.login(this.model.username, this.model.password)
+            .subscribe(
+                data => {
+                    console.log("in routing..");
+                    this.router.navigate([this.returnUrl]);
+                    console.log("routati");
+                },
+                error => {
+                    console.log("errore autenticazione."+error);
+                    this.alertService.error(error);
+                    this.loading = false;
+                });
+        //localStorage.setItem('isLoggedin', 'true');
+    }
 }
