@@ -1,14 +1,14 @@
 import { NgModule } from '@angular/core';
-import { RouterModule } from '@angular/router';
+import { CommonModule } from '@angular/common';
 
-import { NotFoundComponent } from './not-found.component';
 import { NotFoundRoutingModule } from './not-found-routing.module';
+import { NotFoundComponent } from './not-found.component';
 
 @NgModule({
-    imports: [
-        NotFoundRoutingModule,
-        RouterModule
-    ],
-    declarations: [NotFoundComponent]
+  imports: [
+    CommonModule,
+    NotFoundRoutingModule
+  ],
+  declarations: [NotFoundComponent]
 })
-export class NotFoundModule {}
+export class NotFoundModule { }
