@@ -4,11 +4,10 @@ import { Router } from '@angular/router';
 
 @Injectable()
 export class AuthGuard implements CanActivate {
-
-    constructor(private router: Router) { }
+    constructor(private router: Router) {}
 
     canActivate() {
-        if (localStorage.getItem('isLoggedin')) {
+        if (localStorage.getItem('currentUser')) {
             return true;
         }
 
