@@ -1,4 +1,4 @@
-export * from './pipes/shared-pipes.module';
-export * from './components';
 export * from './modules';
-export * from './guard/auth.guard';
+export * from './pipes/shared-pipes.module';
+export * from './modules';
+export * from './guard';
