@@ -1,2 +1,7 @@
-export * from './stat/stat.module';
 export * from './page-header/page-header.module';
+export * from './stat/stat.module';
+export * from './info-aggregate/info-aggregate.module';
+export * from './rich-ass/rich-ass.module';
+export * from './eventi-rich-ass/eventi-rich-ass.module';
+export * from './mezzi-in-servizio/mezzi-in-servizio.module';
+export * from './composizione-partenza-shared/composizione-partenza-shared.module';
