@@ -1,11 +1,15 @@
-import { Component, Input } from '@angular/core';
+import { Component, OnInit, Input } from '@angular/core';
+import { RouterModule } from '@angular/router';
 
 @Component({
     selector: 'app-page-header',
     templateUrl: './page-header.component.html',
     styleUrls: ['./page-header.component.scss']
 })
-export class PageHeaderComponent {
+export class PageHeaderComponent implements OnInit {
     @Input() heading: string;
     @Input() icon: string;
+    constructor() {}
+
+    ngOnInit() {}
 }
