@@ -5,11 +5,8 @@ import { RouterModule } from '@angular/router';
 import { PageHeaderComponent } from './page-header.component';
 
 @NgModule({
-    imports: [
-        CommonModule,
-        RouterModule
-    ],
+    imports: [CommonModule, RouterModule],
     declarations: [PageHeaderComponent],
     exports: [PageHeaderComponent]
 })
-export class PageHeaderModule { }
+export class PageHeaderModule {}
