@@ -3,10 +3,8 @@ import { CommonModule } from '@angular/common';
 import { StatComponent } from './stat.component';
 
 @NgModule({
-    imports: [
-        CommonModule
-    ],
+    imports: [CommonModule],
     declarations: [StatComponent],
     exports: [StatComponent]
 })
-export class StatModule { }
+export class StatModule {}
