@@ -1,10 +1,13 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
+import { FriendlyDatePipe } from "./friendly-date.pipe";
+import { FriendlyHourPipe } from "./friendly-hour.pipe";
 
 @NgModule({
     imports: [
         CommonModule
     ],
-    declarations: []
+    declarations: [FriendlyDatePipe, FriendlyHourPipe],
+    exports: [FriendlyDatePipe, FriendlyHourPipe]
 })
 export class SharedPipesModule { }
