@@ -3,11 +3,14 @@ import { Routes, RouterModule } from '@angular/router';
 import { SignupComponent } from './signup.component';
 
 const routes: Routes = [
-    { path: '', component: SignupComponent }
+    {
+        path: '', component: SignupComponent
+    }
 ];
 
 @NgModule({
-  imports: [RouterModule.forChild(routes)],
-  exports: [RouterModule]
+    imports: [RouterModule.forChild(routes)],
+    exports: [RouterModule]
 })
-export class SignupRoutingModule { }
+export class SignupRoutingModule {
+}
