@@ -1,13 +1,14 @@
 import { Component, OnInit } from '@angular/core';
+import { routerTransition } from '../router.animations';
 
 @Component({
     selector: 'app-signup',
     templateUrl: './signup.component.html',
-    styleUrls: ['./signup.component.scss']
+    styleUrls: ['./signup.component.scss'],
+    animations: [routerTransition()]
 })
 export class SignupComponent implements OnInit {
+    constructor() {}
 
-    constructor() { }
-
-    ngOnInit() { }
+    ngOnInit() {}
 }
